refactor: 공지사항 관련 코드 정리 및 구조 변경
