+++ src/main/java/com/recipe/app/src/common/domain/Notice.java
@@ -1,4 +1,4 @@
-package com.recipe.app.src.notice.models;
+package com.recipe.app.src.common.domain;
 
 import com.recipe.app.common.entity.BaseEntity;
 import lombok.AccessLevel;
@@ -8,13 +8,13 @@ import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 
-@NoArgsConstructor(access = AccessLevel.PUBLIC) // Unit Test 를 위해 PUBLIC
+@NoArgsConstructor(access = AccessLevel.PUBLIC)
 @EqualsAndHashCode(callSuper = false)
-@Data // from lombok
-@Entity // 필수, Class 를 Database Table화 해주는 것이다
-@Table(name = "Notice") // Table 이름을 명시해주지 않으면 class 이름을 Table 이름으로 대체한다.
+@Data
+@Entity
+@Table(name = "Notice")
 public class Notice extends BaseEntity {
-    @Id // PK를 의미하는 어노테이션
+    @Id
     @Column(name = "idx", nullable = false, updatable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Integer idx; // Integer -> Long 21억 지나면 터지니까 Long
