+++ src/main/java/com/recipe/app/src/common/mapper/NoticeRepository.java
@@ -1,6 +1,6 @@
-package com.recipe.app.src.notice;
+package com.recipe.app.src.common.mapper;
 
-import com.recipe.app.src.notice.models.Notice;
+import com.recipe.app.src.common.domain.Notice;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,5 +8,5 @@ import java.util.Optional;
 
 @Repository
 public interface NoticeRepository extends CrudRepository<Notice, Integer> {
-    Notice findFirstByActiveYn(String activeYn);
+    Optional<Notice> findFirstByActiveYn(String activeYn);
 }
