--- src/main/java/com/recipe/app/src/notice/models/GetNoticeRes.java
@@ -1,12 +1,0 @@
-package com.recipe.app.src.notice.models;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class GetNoticeRes {
-    private final int idx;
-    private final String title;
-    private final String content;
-}
